+++ test/envtest/kongstate_consumer_failures_test.go
@@ -15,6 +15,7 @@ import (
 	ctrlclient "sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/kong/kubernetes-ingress-controller/v3/internal/annotations"
+	"github.com/kong/kubernetes-ingress-controller/v3/internal/labels"
 	kongv1 "github.com/kong/kubernetes-ingress-controller/v3/pkg/apis/configuration/v1"
 )
 
@@ -41,21 +42,24 @@ func TestKongStateFillConsumersAndCredentialsFailure(t *testing.T) {
 			ObjectMeta: metav1.ObjectMeta{
 				Name:      "key-auth-cred",
 				Namespace: ns.Name,
+				Labels: map[string]string{
+					labels.CredentialTypeLabel: "key-auth",
+				},
 			},
 			Data: map[string][]byte{
-				"kongCredType": []byte("key-auth"),
-				"key":          []byte("whatever"),
-				"ttl":          []byte("1024"),
+				"key": []byte("whatever"),
+				"ttl": []byte("1024"),
 			},
 		},
 		{
 			ObjectMeta: metav1.ObjectMeta{
 				Name:      "empty-cred",
 				Namespace: ns.Name,
+				Labels: map[string]string{
+					labels.CredentialTypeLabel: "key-auth",
+				},
 			},
-			Data: map[string][]byte{
-				"kongCredType": []byte("key-auth"),
-			},
+			Data: map[string][]byte{},
 		},
 	}
 	for _, secret := range secrets {